jbb/requirements.txt
# On the cluster: packages below are installed via jbb/container/dockerfile-jbb.
# Locally:
#   pip install -r requirements.txt

accelerate>=0.30.0
transformers>=4.36.0,<5
hydra-core>=1.3.0
omegaconf>=2.3.0
loguru>=0.7.0
openai>=1.30.0
